Commenting and cleaning
