+++ src/Examples/CSharp DotNetCore/Program.cs
@@ -6,19 +6,15 @@ namespace CSharpDotNetCore
     {
         static void Main(string[] args)
         {
-            //TestDatatypes.Run();
-            //ExampleGenericTag.UDT_Array();
-            //ExampleAsync.SyncAsyncMultipleTagComparison();
-            //ExampleAsync.AsyncParallelCancellation();
-            //ExampleGenericTag.Run();
-            //ExampleRW.Run();
+            TestDatatypes.Run();
+            ExampleAsync.SyncAsyncMultipleTagComparison();
+            ExampleAsync.AsyncParallelCancellation();
             ExampleArray.Run();
-            //ExampleListTags.Run();
-            //ExampleRW.Run();
-            //ExampleArray.Run();
-            //NativeImportExample.Run();
-            //NativeImportExample.RunCallbackExample();
-            //NativeImportExample.RunLoggerExample();
+            ExampleListTags.Run();
+            ExampleArray.Run();
+            NativeImportExample.Run();
+            NativeImportExample.RunCallbackExample();
+            NativeImportExample.RunLoggerExample();
             Console.ReadKey();
         }
     }
